moved buttons into separate component. Conditionally rendered to have same button for checkin/checkout
